Remove external files
