+++ src/config/segmentation_config.py
@@ -1,5 +1,7 @@
 
 class SegmentationConfig:
-    pass
+    MIN_CHAR_SIZE = (5, 15)
+    MIN_WORD_GAP = 5
+    MIN_LINE_GAP = 5
 
 segmentation_config = SegmentationConfig()
